# [WIP] One hot encoder now errors on any unknown categorical feature.