See the new test case added which fails on the current master.