@amueller 2015-09-24
Sorry still home sick, and the release is a priority atm.
Can you benchmark your changes please?