@vighneshbirodkar 2015-09-28
@amueller You mean these changes themselves ? Or this against the alternative implementation I talked to you about ?