@amueller 2015-10-13
I meant benchmarking this against master. Sorry for the slow turn-around.