@vighneshbirodkar 2015-10-15
Times for fit and predict respectively for 10000 rows and 1000 cols with each column having 100 unique features
#### Master
```
CPU times: user 978 ms, sys: 263 ms, total: 1.24 s
Wall time: 1.24 s
CPU times: user 1.29 s, sys: 544 ms, total: 1.83 s
Wall time: 1.84 s
```
#### This branch
```
CPU times: user 868 ms, sys: 420 ms, total: 1.29 s
Wall time: 3.76 s
CPU times: user 1.32 s, sys: 648 ms, total: 1.97 s
Wall time: 1.97 s
```