@amueller 2015-10-15
So the first is fit and the second predict? It's always good to post the benchmark script, here or as a gist.
The wall time are very different but the cpu time are not. Can you rerun?