@vighneshbirodkar 2015-10-15
```python
from sklearn.preprocessing.data import OneHotEncoder
import numpy as np
import time

ROWS = 10000
COLS = 1000
enc = OneHotEncoder(handle_unknown='error')
data = np.random.randint(1,100,size=(ROWS, COLS))


%time enc.fit(data)
%time enc.transform(data)
```

#### Master
```
CPU times: user 890 ms, sys: 373 ms, total: 1.26 s
Wall time: 3.21 s
CPU times: user 1.33 s, sys: 392 ms, total: 1.72 s
Wall time: 1.72 s
```

#### This branch
```
CPU times: user 877 ms, sys: 327 ms, total: 1.2 s
Wall time: 2 s
CPU times: user 1.3 s, sys: 399 ms, total: 1.7 s
Wall time: 1.7 s
```