@vighneshbirodkar 2015-10-15
@amueller The Wall times vary significantly, I can't seem to figure out why