@amueller 2015-10-16
That's probably because you are running another process? Can you use ``timeit``?