@vighneshbirodkar 2015-10-16
```python
from sklearn.preprocessing.data import OneHotEncoder
import numpy as np
import time

ROWS = 10000
COLS = 1000
enc = OneHotEncoder(handle_unknown='error')
data = np.random.randint(1,100,size=(ROWS, COLS))


%timeit -n 5 -p 5 enc.fit(data)

%timeit -n 5 -p 5 enc.transform(data)
```
#### Master
```
5 loops, best of 3: 1.1307 s per loop
5 loops, best of 3: 1.3384 s per loop
```
#### This Branch
```
5 loops, best of 3: 1.2471 s per loop
5 loops, best of 3: 1.2761 s per loop
```

There is really no reason for a speedup on this branch, I guess the times are really similar.