@MechCoder 2015-10-22
@amueller Is the check `np.any(X < 0)` because of the current logic in the OneHotEncoder code? :P