@amueller 2015-10-29
@MechCoder well because of the current semantics...