@amueller 2015-11-02
I'll try to review soon. Sorry, I'm a bit busy with release work at the moment.