@vighneshbirodkar 2016-02-02
@amueller 
Because we can encounter unknown features even during `fit`, I have moved the validation of `handle_unknown` inside fit