@vighneshbirodkar 2016-02-09
@amueller 
There is an old saying in my native language which roughly translates to "You searched the entire village for a vessel which was under your hand the whole time". The code we needed was lying just a few lines below, I should have noticed it before.