@amueller 2016-02-22
I just realized that we kinda need to break backward compatibility if we want strings to work with this.... meh.