@amueller 2016-02-22
Ok either merge this or deprecate OneHotEncoder and implement it more reasonably ;)