@vighneshbirodkar 2016-02-22
I am in favor of a new implementation.