@vighneshbirodkar 2016-03-12
@amueller @MechCoder 
I wanted to know your opinion about this. Instead of creating a new class I have tried to modify OneHotEncoder and make it backward compatible. I had to remove some attributes, but it supports the `n_values` parameter in a backward compatible manner.